Push CommonSocket members to ClientSocket
